requirements.txt
transformers==4.28.1
torch==1.13.1
tqdm==4.64.0

neo4j==4.4.3

streamlit==1.10.0

requests==2.28.1

pandas==1.4.3

fastapi==0.78.0

uvicorn==0.18.2

localtunnel==2.0.0

numpy==1.22.4

python-dotenv==0.20.0

pydantic==1.9.1